minor updates
bug fixes
